+++ src/LSST/M1M3/TS/Events/ErrorCode.h
@@ -37,7 +37,14 @@ namespace Events {
 
 class ErrorCode : MTM1M3TS_logevent_errorCodeC, public cRIO::Singleton<ErrorCode> {
 public:
-    enum Type { NoFault = 0, EGWPump = 1, TemperatureSensors = 2, EGWPumpStartup = 3 };
+    enum Type {
+        NoFault = 0,
+        EGWPump = 1,
+        TemperatureSensors = 2,
+        EGWPumpStartup = 3,
+        MixingValveTimeout = 4,
+        MixingValveMovedOutOfTarget = 5
+    };
 
     ErrorCode(token);