+++ src/LSST/M1M3/TS/Events/SummaryState.cpp
@@ -58,12 +58,17 @@ bool SummaryState::enabled() {
 void SummaryState::fail(int error_code, const std::string &error_report, const std::string &traceback) {
     _switch_state(MTM1M3TS_shared_SummaryStates_FaultState);
     Events::ErrorCode::instance().set(error_code, error_report, traceback);
+    SPDLOG_ERROR("Faulted ({}): {}", error_code, error_report);
 }
 
 void SummaryState::_switch_state(int new_state) {
     std::lock_guard<std::mutex> lockG(_state_mutex);
     SPDLOG_TRACE("SummaryState::_switch_state from {} to {}", summaryState, new_state);
     if (summaryState == new_state) {
+        // it is legal to go to fault from fault state, as the code might hit
+        if (new_state == MTM1M3TS_shared_SummaryStates_FaultState) {
+            return;
+        }
         throw std::runtime_error(fmt::format("Already in {} state", new_state));
     }
     switch (new_state) {