+++ src/LSST/M1M3/TS/Settings/MixingValve.h
@@ -47,9 +47,17 @@ public:
 
     float percents_to_commanded(float target);
 
+    float current_to_voltage(float current);
+
     float position_to_percents(float position);
 
     PID::PIDParameters pid_parameters;
+
+    // TODO: move to XML - OSW-1311
+    float inPosition;
+    float backlashStep;
+    float minimalMove;
+    float maxMovingTime;
 };
 
 }  // namespace Settings