+++ src/LSST/M1M3/TS/Settings/Setpoint.cpp
@@ -36,8 +36,9 @@ Setpoint::Setpoint(token) {
 }
 
 void Setpoint::load(YAML::Node doc) {
-    SPDLOG_INFO("Loading mixing valve settigns");
+    SPDLOG_INFO("Loading mixing valve settings");
     timestep = doc["Timestep"].as<float>();
+    mixingValveStep = doc["MixingValveStep"].as<float>();
 
     glycolSupplyPercentage = doc["GlycolSupplyPercentage"].as<float>();
     if (glycolSupplyPercentage < 0 || glycolSupplyPercentage > 100) {