+++ src/LSST/M1M3/TS/Settings/Setpoint.h
@@ -43,6 +43,7 @@ public:
     void load(YAML::Node doc);
 
     float timestep;
+    float mixingValveStep;
     float glycolSupplyPercentage;
     float high;
     float low;