+++ src/LSST/M1M3/TS/Telemetry/FinerControl.cpp
@@ -1,0 +1,147 @@
+/*
+ * Task controlling mixing valve position.
+ *
+ * Developed for the Vera C. Rubin Observatory Telescope & Site Software
+ * Systems. This product includes software developed by the Vera C.Rubin
+ * Observatory Project (https://www.lsst.org). See the COPYRIGHT file at the
+ * top-level directory of this distribution for details of code ownership.
+ *
+ * This program is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation, either version 3 of the License, or (at your option)
+ * any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include "Events/ErrorCode.h"
+#include "Events/SummaryState.h"
+#include "Settings/MixingValve.h"
+#include "Telemetry/FinerControl.h"
+#include "Telemetry/MixingValve.h"
+
+using namespace LSST::M1M3::TS::Telemetry;
+
+FinerControl::FinerControl(token) {
+    _move_timeout =
+            std::chrono::steady_clock::now() +
+            2 * std::chrono::milliseconds((int)(Settings::MixingValve::instance().maxMovingTime * 1000));
+    _comp_setpoint = NAN;
+    _last_setpoint = 0;
+    state = ON_TARGET;
+}
+
+void FinerControl::set_target(float demand) {
+    std::lock_guard<std::mutex> lock_g(_lock);
+
+    auto backlash_step = Settings::MixingValve::instance().backlashStep;
+
+    if (demand != _last_setpoint) {
+        if (abs(demand - _last_setpoint) > Settings::MixingValve::instance().minimalMove) {
+            state = MOVING_TO_TARGET;
+        } else {
+            state = MOVING_TO_COMPENSATED_TARGET;
+            // The following code handles all cases of future demand
+            // (_comp_setpoint) being < 0 or > 100.
+            //
+            // First if demand is lower than the current setpoint..
+            if (demand < _last_setpoint) {
+                // if demand is below backlash_step, move to higher opening
+                // first - set _comp_setpoint to _last_setpoint + backlash_step
+                if (demand < backlash_step) {
+                    _comp_setpoint = _last_setpoint + backlash_step;
+                    // in other cases, move down first - demand will then move up
+                } else {
+                    _comp_setpoint = demand - backlash_step;
+                }
+                // Similar for when moving up - don't move over 100, and guarantee
+                // that the minimal move is larger than backlash_step.
+            } else {
+                if (demand > (100 - backlash_step)) {
+                    _comp_setpoint = _last_setpoint - backlash_step;
+                } else {
+                    _comp_setpoint = demand + backlash_step;
+                }
+            }
+        }
+        // move to demand after mixing valve moves close enough to
+        // _comp_setpoint target.
+        _last_setpoint = demand;
+        _move_timeout =
+                std::chrono::steady_clock::now() +
+                std::chrono::milliseconds((int)(Settings::MixingValve::instance().maxMovingTime * 1000));
+    }
+}
+
+float FinerControl::get_target(float valve_position) {
+    std::lock_guard<std::mutex> lock_g(_lock);
+
+    auto now = std::chrono::steady_clock::now();
+
+    auto &mixing_settings = Settings::MixingValve::instance();
+
+    // do not change state for at least two seconds, as the valve moves with delay
+    bool transition =
+            (_move_timeout - now) < std::chrono::milliseconds((int)mixing_settings.maxMovingTime * 800);
+
+    /**
+     * Runs state machine, based on what the valve shall do.
+     */
+    switch (state) {
+        // when moving to compensated target, check if mixing valve is close
+        // enough to it. If that's the case, change state to MOVING_TO_TARGET.
+        case MOVING_TO_COMPENSATED_TARGET:
+            if (abs(valve_position - _comp_setpoint) < mixing_settings.inPosition && transition) {
+                state = MOVING_TO_TARGET;
+                _move_timeout = now + std::chrono::milliseconds((int)(mixing_settings.maxMovingTime * 1000));
+                return _last_setpoint;
+            }
+            if (now >= _move_timeout) {
+                state = FAULTED;
+                Events::SummaryState::instance().fail(
+                        Events::ErrorCode::MixingValveTimeout,
+                        fmt::format("Timeout moving to compensated target value "
+                                    "(valve_position is {}, demand is {}).",
+                                    valve_position, _comp_setpoint),
+                        "");
+                return NAN;
+            }
+            return _comp_setpoint;
+        // when moving to target
+        case MOVING_TO_TARGET:
+            if (abs(valve_position - _last_setpoint) < mixing_settings.inPosition && transition) {
+                state = ON_TARGET;
+                return NAN;
+            } else if (now >= _move_timeout) {
+                Events::SummaryState::instance().fail(
+                        Events::ErrorCode::MixingValveTimeout,
+                        fmt::format("Timeout moving to target value (valve_position is {}, demand is {}).",
+                                    valve_position, _last_setpoint),
+                        "");
+                state = FAULTED;
+                return NAN;
+            }
+            return _last_setpoint;
+        case ON_TARGET:
+            if (abs(valve_position - _last_setpoint) >= mixing_settings.inPosition) {
+                Events::SummaryState::instance().fail(
+                        Events::ErrorCode::MixingValveMovedOutOfTarget,
+                        fmt::format("Moved out of target while on target: {}, demand was {}.", valve_position,
+                                    _last_setpoint),
+                        "");
+                state = FAULTED;
+                return NAN;
+            }
+            return NAN;
+        case FAULTED:
+            return NAN;
+        default:
+            return _last_setpoint;
+    }
+}