+++ src/ts-M1M3thermald.cpp
@@ -84,7 +84,13 @@ void M1M3thermald::init() {
     LSST::cRIO::Settings::Path::setRootPath(getConfigRoot());
 
     SPDLOG_INFO("Initializing M1M3TS SAL");
-    _m1m3tsSAL = std::make_shared<SAL_MTM1M3TS>();
+    try {
+        _m1m3tsSAL = std::make_shared<SAL_MTM1M3TS>();
+    } catch (std::runtime_error &er) {
+        SPDLOG_CRITICAL("Cannot initialize SAL: {}", er.what());
+        throw er;
+    }
+
     _m1m3tsSAL->setDebugLevel(getDebugLevelSAL());
 
     addSink(std::make_shared<SALSink_mt>(_m1m3tsSAL));